requirements.txt
django==3.1.7
gunicorn==20.1.0
django-heroku==0.3.1
djangorestframework==3.12.4